re-factored web server templates, something is off
