+++ src/Main.cpp
@@ -178,13 +178,13 @@ void setup() {
     //enqueues the alarm setup event
     enqueueAlarmSetup();
 
+    //wait for the other core to finish all initializations before allowing web server to respond to requests
+    ulTaskNotifyTake(pdTRUE, portMAX_DELAY);
     web::server_setup();
 
     watchdogSetup();
     Log.info(F("Main Core 0 Setup completed, CORE1 notified of WiFi %d. System status: %#hX"), c1NtfStatus, sysInfo->getSysStatus());
     logSystemInfo();
-    //wait for the other core to finish all initializations before allowing web server to respond to requests
-    ulTaskNotifyTake(pdTRUE, portMAX_DELAY);
 }
 
 /**
